Disable CheckSpace in default config
Disable CheckSpace in the default pacman.conf because on many MSYS2
installations the free space checking is excruciatingly slow.

The extra safety afforded by this feature is not justified by the very
painful experience of many users, and the situation of not having
sufficient free space for packages is very unlikely on modern hardware.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>